+++ src/pages/app.jsx
@@ -23,8 +23,10 @@ export default function App() {
         <Route exact path="/weather" element={<Weather />} />
         <Route exact path="/advancedWeather" element={<AdvancedWeather />} />
         <Route exact path="/weather/:city" element={<GetOpenWeatherData />} /> (// changes url according to city)
-        <Route exact path="/weather/:countryCode/:city" element={<GetOpenWeatherData />} />
+        <Route exact path="/weatherCountry/:countryCode/:city" element={<GetOpenWeatherData />} />
+        <Route exact path="/weatherLocation/:latitude/:longitude" element={<GetOpenWeatherData/>} />
         <Route exact path="/weather/:city/:lat/:lon" element={<MultipleWeatherData />} />
+
       </Routes>
     </Router>
   );